Fixed contract screen
