+++ Projects/Claims/Sources/Journeys/ClaimJourneys.swift
@@ -71,19 +71,7 @@ public class ClaimJourneys {
     private static func openSelectContractScreen() -> some JourneyPresentation {
         HostingJourney(
             SubmitClaimStore.self,
-            rootView: CheckboxPickerScreen<FlowClaimContractSelectOptionModel>(
-                items: {
-                    let store: SubmitClaimStore = globalPresentableStoreContainer.get()
-                    return store.state.contractStep?.availableContractOptions
-                        .compactMap({ (object: $0, displayName: $0.displayName) }) ?? []
-                }(),
-                preSelectedItems: { [] },
-                onSelected: { selectedContract in
-                    let store: SubmitClaimStore = globalPresentableStoreContainer.get()
-                    store.send(.contractSelectRequest(contractId: selectedContract.first?.id))
-                },
-                singleSelect: true
-            )
+            rootView: SelectContractScreen()
         ) {
             action in
             getScreen(for: action)
